+++ apps/api/src/services/data-provider/ghostfolio-scraper-api/interfaces/scraper-configuration.interface.ts
@@ -1,0 +1,4 @@
+export interface ScraperConfiguration {
+  selector: string;
+  url: string;
+}